requirements.txt
streamlit
tensorflow
numpy
opencv-python
pillow
scikit-learn
joblib
scikit-image
gdown